moving data grid props around
