+++ src/components/datagrid/column_sorting_draggable.tsx
@@ -6,34 +6,17 @@
  * Side Public License, v 1.
  */
 
-import React, { FunctionComponent, ReactChild } from 'react';
-import { EuiI18n } from '../i18n';
-import { EuiDraggable } from '../drag_and_drop';
+import React, { FunctionComponent } from 'react';
 import { EuiScreenReaderOnly } from '../accessibility';
+import { EuiButtonGroup, EuiButtonIcon } from '../button';
+import { EuiDraggable } from '../drag_and_drop';
 import { EuiFlexGroup, EuiFlexItem } from '../flex';
-import { EuiButtonIcon, EuiButtonGroup } from '../button';
+import { EuiI18n } from '../i18n';
 import { EuiIcon } from '../icon';
 import { EuiText } from '../text';
-import {
-  getDetailsForSchema,
-  EuiDataGridSchema,
-  EuiDataGridSchemaDetector,
-} from './data_grid_schema';
-import { EuiDataGridSorting } from './data_grid_types';
 import { EuiToken } from '../token';
-
-export interface EuiDataGridColumnSortingDraggableProps {
-  id: string;
-  direction: string;
-  index: number;
-  sorting: EuiDataGridSorting;
-  schema: EuiDataGridSchema;
-  schemaDetectors: EuiDataGridSchemaDetector[];
-  /**
-   * Value to be shown in column sorting popover.
-   */
-  display: string;
-}
+import { getDetailsForSchema } from './data_grid_schema';
+import { EuiDataGridColumnSortingDraggableProps } from './data_grid_types';
 
 export const defaultSortAscLabel = (
   <EuiI18n token="euiColumnSortingDraggable.defaultSortAsc" default="A-Z" />
@@ -86,17 +69,12 @@ export const EuiDataGridColumnSortingDraggable: FunctionComponent<EuiDataGridCol
             state.isDragging && 'euiDataGridColumnSorting__item-isDragging'
           }`}>
           <EuiScreenReaderOnly>
-            <p>
-              <EuiI18n
-                token="euiColumnSortingDraggable.activeSortLabel"
-                default="is sorting this data grid">
-                {(activeSortLabel: ReactChild) => (
-                  <span>
-                    {display} {activeSortLabel}
-                  </span>
-                )}
-              </EuiI18n>
-            </p>
+            <EuiI18n
+              token="euiColumnSortingDraggable.activeSortLabel"
+              default="{display} is sorting this data grid"
+              values={{ display }}>
+              {(activeSortLabel: string) => <p>{activeSortLabel}</p>}
+            </EuiI18n>
           </EuiScreenReaderOnly>
           <EuiFlexGroup
             gutterSize="xs"
@@ -106,12 +84,13 @@ export const EuiDataGridColumnSortingDraggable: FunctionComponent<EuiDataGridCol
             <EuiFlexItem grow={false}>
               <EuiI18n
                 token="euiColumnSortingDraggable.removeSortLabel"
-                default="Remove from data grid sort:">
-                {(removeSortLabel: ReactChild) => (
+                default="Remove {display} from data grid sort"
+                values={{ display }}>
+                {(removeSortLabel: string) => (
                   <EuiButtonIcon
                     color="text"
                     className="euiDataGridColumnSorting__button"
-                    aria-label={`${removeSortLabel} ${id}`}
+                    aria-label={removeSortLabel}
                     iconType="cross"
                     onClick={() => {
                       const nextColumns = [...sorting.columns];
@@ -142,10 +121,11 @@ export const EuiDataGridColumnSortingDraggable: FunctionComponent<EuiDataGridCol
             <EuiFlexItem className="euiDataGridColumnSorting__orderButtons">
               <EuiI18n
                 token="euiColumnSortingDraggable.toggleLegend"
-                default="Select sorting method for field: ">
-                {(toggleLegend: ReactChild) => (
+                default="Select sorting method for {display}"
+                values={{ display }}>
+                {(toggleLegend: string) => (
                   <EuiButtonGroup
-                    legend={`${toggleLegend} ${id}`}
+                    legend={toggleLegend}
                     name={id}
                     isFullWidth
                     options={toggleOptions}
