+++ src/components/datagrid/data_grid_body.tsx
@@ -6,85 +6,44 @@
  * Side Public License, v 1.
  */
 
+import classNames from 'classnames';
 import React, {
   forwardRef,
   FunctionComponent,
+  useCallback,
+  useContext,
   useEffect,
   useMemo,
   useRef,
-  useCallback,
-  useContext,
   useState,
 } from 'react';
-import classNames from 'classnames';
-import tabbable from 'tabbable';
 import {
   GridChildComponentProps,
   VariableSizeGrid as Grid,
   VariableSizeGridProps,
 } from 'react-window';
+import tabbable from 'tabbable';
 import { EuiCodeBlock } from '../code';
-import {
-  EuiDataGridControlColumn,
-  EuiDataGridColumn,
-  EuiDataGridColumnWidths,
-  EuiDataGridPopoverContents,
-  EuiDataGridInMemory,
-  EuiDataGridInMemoryValues,
-  EuiDataGridPaginationProps,
-  EuiDataGridPopoverContent,
-  EuiDataGridRowHeightsOptions,
-  EuiDataGridStyle,
-} from './data_grid_types';
-import { EuiDataGridCell, EuiDataGridCellProps } from './data_grid_cell';
-import {
-  EuiDataGridSchema,
-  EuiDataGridSchemaDetector,
-} from './data_grid_schema';
-import { EuiDataGridFooterRow } from './data_grid_footer_row';
-import {
-  EuiDataGridHeaderRow,
-  EuiDataGridHeaderRowProps,
-} from './data_grid_header_row';
 import {
   EuiMutationObserver,
   useMutationObserver,
 } from '../observer/mutation_observer';
+import { useResizeObserver } from '../observer/resize_observer';
 import { EuiText } from '../text';
+import { EuiDataGridCell } from './data_grid_cell';
 import {
   DataGridSortingContext,
   DataGridWrapperRowsContext,
 } from './data_grid_context';
-import { useResizeObserver } from '../observer/resize_observer';
-import { RowHeightUtils } from './row_height_utils';
-
-export interface EuiDataGridBodyProps {
-  isFullScreen: boolean;
-  columnWidths: EuiDataGridColumnWidths;
-  defaultColumnWidth?: number | null;
-  leadingControlColumns?: EuiDataGridControlColumn[];
-  trailingControlColumns?: EuiDataGridControlColumn[];
-  columns: EuiDataGridColumn[];
-  schema: EuiDataGridSchema;
-  schemaDetectors: EuiDataGridSchemaDetector[];
-  popoverContents?: EuiDataGridPopoverContents;
-  rowCount: number;
-  renderCellValue: EuiDataGridCellProps['renderCellValue'];
-  renderFooterCellValue?: EuiDataGridCellProps['renderCellValue'];
-  inMemory?: EuiDataGridInMemory;
-  inMemoryValues: EuiDataGridInMemoryValues;
-  interactiveCellId: EuiDataGridCellProps['interactiveCellId'];
-  pagination?: EuiDataGridPaginationProps;
-  setColumnWidth: (columnId: string, width: number) => void;
-  headerIsInteractive: boolean;
-  handleHeaderMutation: MutationCallback;
-  setVisibleColumns: EuiDataGridHeaderRowProps['setVisibleColumns'];
-  switchColumnPos: EuiDataGridHeaderRowProps['switchColumnPos'];
-  toolbarHeight: number;
-  rowHeightsOptions?: EuiDataGridRowHeightsOptions;
-  rowHeightUtils: RowHeightUtils;
-  gridStyles?: EuiDataGridStyle;
-}
+import { EuiDataGridFooterRow } from './data_grid_footer_row';
+import { EuiDataGridHeaderRow } from './data_grid_header_row';
+import {
+  EuiDataGridBodyProps,
+  EuiDataGridInMemoryValues,
+  EuiDataGridPopoverContent,
+  EuiDataGridPopoverContents,
+  EuiDataGridSchemaDetector,
+} from './data_grid_types';
 
 export const VIRTUALIZED_CONTAINER_CLASS = 'euiDataGrid__virtualized';
