+++ src/components/datagrid/data_grid_context.tsx
@@ -6,16 +6,12 @@
  * Side Public License, v 1.
  */
 
-import React, { ReactElement } from 'react';
-import { EuiDataGridFocusedCell, EuiDataGridSorting } from './data_grid_types';
-
-export interface DataGridFocusContextShape {
-  setFocusedCell: (cell: EuiDataGridFocusedCell) => void;
-  onFocusUpdate: (
-    cell: EuiDataGridFocusedCell,
-    updateFocus: Function
-  ) => () => void;
-}
+import React from 'react';
+import {
+  DataGridFocusContextShape,
+  DataGridWrapperRowsContentsShape,
+  EuiDataGridSorting,
+} from './data_grid_types';
 
 export const DataGridFocusContext = React.createContext<
   DataGridFocusContextShape
@@ -28,11 +24,6 @@ export const DataGridSortingContext = React.createContext<
   EuiDataGridSorting | undefined
 >(undefined);
 
-export interface DataGridWrapperRowsContentsShape {
-  headerRowHeight: number;
-  headerRow: ReactElement;
-  footerRow: ReactElement | null;
-}
 export const DataGridWrapperRowsContext = React.createContext<
   DataGridWrapperRowsContentsShape
 >({ headerRow: <div />, headerRowHeight: 0, footerRow: null });
