+++ src/components/datagrid/data_grid_inmemory_renderer.tsx
@@ -9,27 +9,18 @@
 import React, {
   FunctionComponent,
   JSXElementConstructor,
-  useEffect,
   useCallback,
+  useEffect,
   useMemo,
   useState,
 } from 'react';
 import { createPortal } from 'react-dom';
-import {
-  EuiDataGridCellValueElementProps,
-  EuiDataGridCellProps,
-} from './data_grid_cell';
-import { EuiDataGridColumn, EuiDataGridInMemory } from './data_grid_types';
 import { enqueueStateChange } from '../../services/react';
 import { EuiMutationObserver } from '../observer/mutation_observer';
-
-export interface EuiDataGridInMemoryRendererProps {
-  inMemory: EuiDataGridInMemory;
-  columns: EuiDataGridColumn[];
-  rowCount: number;
-  renderCellValue: EuiDataGridCellProps['renderCellValue'];
-  onCellRender: (rowIndex: number, columnId: string, value: string) => void;
-}
+import {
+  EuiDataGridCellValueElementProps,
+  EuiDataGridInMemoryRendererProps,
+} from './data_grid_types';
 
 function noop() {}
