+++ src/components/datagrid/data_grid_schema.tsx
@@ -6,60 +6,16 @@
  * Side Public License, v 1.
  */
 
-import React, { useMemo, ReactNode } from 'react';
+import React, { useMemo } from 'react';
+import { EuiI18n } from '../i18n';
 import {
   EuiDataGridColumn,
   EuiDataGridInMemory,
   EuiDataGridInMemoryValues,
+  EuiDataGridSchema,
+  EuiDataGridSchemaDetector,
+  SchemaTypeScore,
 } from './data_grid_types';
-
-import { EuiI18n } from '../i18n';
-
-import { IconType } from '../icon';
-import { EuiTokenProps } from '../token';
-
-export interface EuiDataGridSchemaDetector {
-  /**
-   * The name of this data type, matches #EuiDataGridColumn / schema `schema`
-   */
-  type: string;
-  /**
-   * The function given the text value of a cell and returns a score of [0...1] of how well the value matches this data type
-   */
-  detector: (value: string) => number;
-  /**
-   * A custom comparator function when performing in-memory sorting on this data type, takes `(a: string, b: string, direction: 'asc' | 'desc) => -1 | 0 | 1`
-   */
-  comparator?: (a: string, b: string, direction: 'asc' | 'desc') => -1 | 0 | 1;
-  /**
-   * The icon used to visually represent this data type. Accepts any `EuiIcon IconType`.
-   */
-  icon: IconType;
-  /**
-   * The color associated with this data type; it's used to color the icon token
-   */
-  color?: EuiTokenProps['color'] | string;
-  /**
-   * Text for how to represent an ascending sort of this data type, e.g. 'A -> Z'
-   */
-  sortTextAsc: ReactNode;
-  /**
-   * Text for how to represent a descending sort of this data type, e.g. 'Z -> A'
-   */
-  sortTextDesc: ReactNode;
-  /**
-   * Whether this column is sortable (defaults to true)
-   */
-  isSortable?: boolean;
-  /**
-   *  This property controls the capitalization of text
-   */
-  textTransform?: 'uppercase' | 'lowercase' | 'capitalize';
-  /**
-   * Default sort direction of the column
-   */
-  defaultSortDirection?: 'asc' | 'desc';
-}
 
 const numericChars = new Set([
   '0',
@@ -75,6 +31,7 @@ const numericChars = new Set([
   '.',
   '-',
 ]);
+
 export const schemaDetectors: EuiDataGridSchemaDetector[] = [
   {
     type: 'boolean',
@@ -252,15 +209,6 @@ export const schemaDetectors: EuiDataGridSchemaDetector[] = [
     ),
   },
 ];
-
-export interface EuiDataGridSchema {
-  [columnId: string]: { columnType: string | null };
-}
-
-export interface SchemaTypeScore {
-  type: string;
-  score: number;
-}
 
 function scoreValueBySchemaType(
   value: string,
