+++ src/components/datagrid/index.ts
@@ -6,24 +6,7 @@
  * Side Public License, v 1.
  */
 
-export { EuiDataGridColumnSortingDraggableProps } from './column_sorting_draggable';
-export { EuiDataGrid, EuiDataGridProps } from './data_grid';
-export { EuiDataGridBodyProps } from './data_grid_body';
-export {
-  EuiDataGridCellProps,
-  EuiDataGridCellValueProps,
-  EuiDataGridCellValueElementProps,
-} from './data_grid_cell';
-export { EuiDataGridColumnResizerProps } from './data_grid_column_resizer';
-export { EuiDataGridHeaderRowProps } from './data_grid_header_row';
-export { EuiDataGridHeaderCellProps } from './data_grid_header_cell';
-export { EuiDataGridControlHeaderRowProps } from './data_grid_control_header_cell';
-export { EuiDataGridInMemoryRendererProps } from './data_grid_inmemory_renderer';
-export {
-  EuiDataGridSchema,
-  EuiDataGridSchemaDetector,
-  SchemaTypeScore,
-} from './data_grid_schema';
+export { EuiDataGrid } from './data_grid';
 export { useDataGridColumnSelector } from './column_selector';
 export { useDataGridColumnSorting } from './column_sorting';
 export { useDataGridStyleSelector } from './style_selector';
